Manage desktops using rofi and btops I figured out a solution to the desktop management without having to resort to a bar.
This method utilises a go utility, btops, to dynamically create desktops that provide
a list of client names to use as the desktop names. This allows for an infinite amount
for desktops. I then manage this using a simple rofi mode I scripted.

Only limitation is that it may get difficult to differentiate between similar windows
on different desktops, but that's a minor inconvenience.